requirements.txt
Django>=5.0  # Latest stable version
django-environ>=0.11.2  # Environment variable management
django-extensions>=3.2.3  # Extra Django tools
pillow>=10.2.0  # Image processing

# Database
#psycopg2-binary>=2.9.9  # PostgreSQL
#mysqlclient>=2.2.0  # MySQL

# Permitir LLamadas del Front
django-cors-headers>=4.3.1

# Django REST Framework and JWT Authentication
djangorestframework>=3.14.0  # DRF for API development
djangorestframework-simplejwt>=5.3.0  # JWT authentication

# Testing
pytest
pytest-django

#validation 
python-dotenv
